backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.6.0
beanie==1.26.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3>=1.34.0
slowapi>=0.1.9
fastapi-cache2>=0.2.1
loguru>=0.7.0
google-genai>=1.0.0